backend/requirements.txt
anyio==3.6.2
bleach==6.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
deprecation==2.1.0
docutils==0.20.1
dotty-dict==1.3.1
fastapi==0.95.0
gitdb==4.0.10
GitPython==3.1.32
gotrue==1.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
invoke==1.7.3
jaraco.classes==3.3.0
keyring==24.2.0
more-itertools==9.1.0
packaging==23.1
pkginfo==1.9.6
postgrest==0.10.6
pydantic==1.10.7
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.15.0
python-semantic-release==7.33.2
readme-renderer==40.0
realtime==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.26.1
storage3==0.5.2
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
tomlkit==0.11.8
tqdm==4.65.0
twine==3.8.0
typing_extensions==4.5.0
urllib3==2.0.4
uvicorn==0.21.1
webencodings==0.5.1
websockets==10.4
zipp==3.16.2
